requirements.txt
streamlit==1.35.0
scikit-learn==1.4.2
pandas==2.2.2
pyarrow==14.0.2
numpy==1.26.4
mlflow==2.13.0
dagshub==0.3.25
joblib==1.4.2
python-dotenv==1.0.1
supabase>=2.3.0
duckdb>=0.10.0
dvc>=3.48.0
